+++ CHANGELOG.rst
@@ -12,6 +12,7 @@ Changelog
 * Add new method :meth:`pymaf.MafFrame.plot_clonality`.
 * Fix a bug in :meth:`pymaf.MafFrame.plot_evolution` when there are no variants to display for the specified samples.
 * Add new method :meth:`pymaf.MafFrame.plot_genepair`.
+* Add new method :meth:`pymaf.MafFrame.plot_interactions`.
 
 0.18.0 (2021-07-20)
 -------------------